pyproject.toml
[project]
name = "basic-whisper-mcp-server"
version = "0.1.0"
description = "MCP server for audio transcription using OpenAI Whisper"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.4",
    "openai-whisper",
]

[project.scripts]
basic-whisper-mcp-server = "basic_whisper_mcp_server.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
